first working version of localstorage object
